+++ h3ron/src/algorithm/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "indexmap")]
+pub mod cell_clusters;
+pub mod smoothen;
+
+#[cfg(feature = "indexmap")]
+pub use cell_clusters::*;
+pub use smoothen::*;