+++ h3ron/CHANGES.md
@@ -15,6 +15,7 @@ python extension.
 * Add example describing how to compile to the wasm32-wasi target.
 * Parse `H3Cell` from different string representations using `H3Cell::from_str`. [#52](https://github.com/nmandery/h3ron/pull/52)
 * `ToLine` trait and implementation for `H3DirectedEdge`.
+* Added `h3ron::algorithm::find_cell_clusters` and `h3ron::algorithm::find_cell_clusters_eq_value`.
 
 ### Changed
 * Use only types from `geo-types` in the public API, not from `geo`.